Refactor grid and pixel sizing
Previously, the grid's width was set after determining a grid pixel's
width in order to work around a floating point precision error. This led
to a slightly variable grid width and height.

Now, the grid pixels are placed in rows. Pixels stretch to fill their
rows, and rows stretch to fill the grid. This allow for the grid width
and height to be declared in CSS.